Moved update_wrapper and wraps to functoolz.compatibility.
Feels more comfortable in compatibility rather than in functoolz module since they're backports.
